feat(dashboard): add mobile layout with panel switcher
Below 768px viewport width, replace the resizable sidebar layout with
a bottom tab bar switching between Sessions, Viewport, and Tools panels.

- Extract ViewportPanel from Dashboard for reuse across layouts
- Move browser-scoped state from Dashboard into App (index.tsx)
- Delete dashboard.tsx; split into index.tsx (state) and viewportPanel.tsx (UI)
- Add useIsMobile hook and MobilePanelSwitcher component
- Cross-panel navigation: session select → viewport, pick locator → viewport,
  element picked → tools
- Persist active panel selection across reloads
- Mobile CSS: compact toolbar, touch-friendly close buttons
- Desktop layout unchanged
