+++ packages/dashboard/src/mobilePanelSwitcher.css
@@ -1,0 +1,79 @@
+/**
+ * Copyright (c) Microsoft Corporation.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+.mobile-panel-switcher {
+  display: flex;
+  flex-direction: column;
+  height: 100%;
+  width: 100%;
+  overflow: hidden;
+}
+
+.mobile-panels {
+  flex: 1;
+  min-height: 0;
+  position: relative;
+}
+
+.mobile-panel {
+  flex-direction: column;
+  height: 100%;
+  width: 100%;
+  overflow: hidden;
+}
+
+/* -- Bottom tab bar -- */
+.mobile-tab-bar {
+  display: flex;
+  align-items: stretch;
+  height: 52px;
+  min-height: 52px;
+  background: var(--color-canvas-overlay);
+  border-top: 1px solid var(--color-border-muted);
+  padding: 0;
+  gap: 0;
+}
+
+.mobile-tab-bar-item {
+  flex: 1;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  background: none;
+  border: none;
+  color: var(--color-fg-muted);
+  cursor: pointer;
+  padding: 4px 0;
+  position: relative;
+  font-size: 12px;
+  font-weight: 500;
+  -webkit-tap-highlight-color: transparent;
+}
+
+.mobile-tab-bar-item.active {
+  color: var(--color-accent-fg);
+}
+
+.mobile-tab-bar-item.active::before {
+  content: '';
+  position: absolute;
+  top: 0;
+  left: 25%;
+  right: 25%;
+  height: 2px;
+  background: var(--color-accent-fg);
+  border-radius: 0 0 2px 2px;
+}
